HT-465: Put TestErrorReporter into its own file and dealt with other code-review feedback.
